requirements.txt
discord.py
nltk
pyfiglet
requests
colorama
